+++ src/Regressors/KNNRegressor.php
@@ -256,6 +256,7 @@ class KNNRegressor implements Estimator, Learner, Online, Parallel, Persistable
      */
     public function predictSample(array $sample) : float
     {
+        /** @var list<float> $labels */
         [$labels, $distances] = $this->nearest($sample);
 
         if ($this->weighted) {